requirements.txt
requests>=2.12.4
gdapi-python>=0.5.3
gitpython>=2.1.1
pyaml>=16.12.2
